yolo working...ish
Problems with async. Seems to not pause the method during object detection even though the await keyword should make it pause? Removing async entirely might be the play, and then we just need to find the maximum resolution usable that still allows for real time detection.

Was also getting cuda problems with async, probably because it's trying to run far too many detections at a time. Hopefully this won't be a problem elsewhere